requirements.txt
discord.py
psutil
pydcs
requests